закончить разбивку на слайсы
