+++ src/pages/login/login.tsx
@@ -2,14 +2,14 @@ import Header from '../../components/header/header';
 import SignIn from '../../components/sign-in/sign-in';
 import {Link, useNavigate} from 'react-router-dom';
 import {RoutePath} from '../../routes';
-import {setActiveCityName} from '../../store/action';
 import {useAppDispatch} from '../../store/use-app-dispatch';
 import {getRandomArrayElement} from '../../utils';
 import { CITIES_NAMES} from '../../const';
 import { CityName } from '../../types';
 import {useAppSelector} from '../../store/use-app-selector';
-import {getIsAuthed} from '../../store/selectors';
 import {useEffect} from 'react';
+import {getIsAuthed} from '../../store/slices/user/selectors.ts';
+import {setActiveCityName} from '../../store/slices/app/app.ts';
 
 function Login(): JSX.Element {
   const randomCityName = getRandomArrayElement<CityName>(CITIES_NAMES);
