--- src/store/action.ts
@@ -1,5 +1,0 @@
-import {createAction} from '@reduxjs/toolkit';
-import {CityName, SortType} from '../types';
-
-// export const setActiveCityName = createAction<CityName>('main/setActiveCityName');
-// export const setOfferSortOption = createAction<SortType>('main/changeOfferSortOption');
